# specs/api.yaml
swagger: '2.0'
info:
  description:
  version: 0.0.1-SNAPSHOT
  title: Swagger Employee Demo
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: localhost
basePath: /v1
tags:
  - name: employee
    description: Everything about your Employee
schemes:
  - http #security schemes
paths:
  '/employee':
    get:
      tags:
      - employee
      summary: List employees
      description: Returns list of Employee
      operationId: getAllEmployees
      produces:
      - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Employee'
    post:
      tags:
      - employee
      summary: Create employee
      description: Creates a new Employee
      operationId: createEmployee
      consumes:
      - application/json
      - application/xml
      produces:
      - application/json
      requestBody:
#        description: Optional description in *Markdown*
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
          application/xml:
            schema:
              $ref: '#/components/schemas/Employee'
          text/plain:
            schema:
              type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Employee'
        '400':
          description: Invalid input
  '/employee/{employeeId}':
    get:
      tags:
        - employee
      summary: Find employee by ID
      description: Returns a single Employee
      operationId: getEmployeeDetails
      produces:
        - application/json
      parameters:
        - name: employeeId
          in: path
          description: ID of Employee to return
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Employee'
        '400':
          description: Invalid Employee ID supplied
        '404':
          description: Employee not found
    delete:
      tags:
      - employee
      summary: Delete employee by ID
      description: Deletes a single Employee
      operationId: deleteEmployee
      parameters:
      - name: employeeId
        in: path
        description: ID of Employee to delete
        required: true
        type: integer
        format: int64
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid Employee ID supplied
        '404':
          description: Employee not found
    put:
      tags:
      - employee
      summary: Update employee
      description: Updates an existing Employee
      operationId: updateEmployee
      consumes:
      - application/json
      - application/xml
      produces:
      - application/json
      parameters:
      - name: employeeId
        in: path
        description: ID of Employee to update
        required: true
        type: integer
        format: int64
      requestBody:
        #        description: Optional description in *Markdown*
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
          application/xml:
            schema:
              $ref: '#/components/schemas/Employee'
          text/plain:
            schema:
              type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Employee'
        '400':
          description: Invalid input
        '404':
          description: Employee not found
definitions:
  Employee:
    type: object
    properties:
      id:
        type: integer
        format: int64
      firstName:
        type: string
        description: Employee First Name
      lastName:
        type: string
        description: Employee Last Name
    xml:
      name: employee